+++ app/src/main/java/com/simenko/qmapp/ui/main/structure/forms/line/LineForm.kt
@@ -93,8 +93,8 @@ fun LineForm(
             InfoLine(modifier = modifier.padding(start = 0.dp), title = "Channel", body = concatTwoStrings(lnComplete.channelWithParents.channelAbbr, lnComplete.channelWithParents.channelDesignation))
             Spacer(modifier = Modifier.height(10.dp))
             RecordFieldItem(
-                valueParam = Triple(lnComplete.line.lineOrder.let { if (it == NoRecord.num) EmptyString.str else it }.toString(), fillInErrors.lineOrderError) {
-                    viewModel.setLineOrder(if(it == EmptyString.str) NoRecord.num else it.toInt())
+                valueParam = Triple(lnComplete.line.lineOrder.let { if (it == NoRecord.num.toInt()) EmptyString.str else it }.toString(), fillInErrors.lineOrderError) {
+                    viewModel.setLineOrder(if(it == EmptyString.str) NoRecord.num.toInt() else it.toInt())
                 },
                 keyboardNavigation = Pair(orderFR) { abbreviationFR.requestFocus() },
                 keyBoardTypeAction = Pair(KeyboardType.Number, ImeAction.Next),
